function/requirements.txt
flask >= 1.1.2, < 2.0.0
python-dotenv >= 0.19.1, < 1.0.0
